refactor(app): extract entity-create option loaders into a composable (#346)
Move the inheritance/phenotype/variation/status option-tree loading (the pure
flattenTreeOptions/transformModifierTree transforms plus the typed-list API
loaders) out of CreateEntity.vue into useEntityCreateOptions.ts, with unit tests
for the pure transforms. Behavior-preserving; CreateEntity.vue 637 -> 540
(under the 600-line ceiling).

Part of #398, #346.
