+++ scripts/code-quality-file-size-baseline.tsv
@@ -54,7 +54,7 @@ app/src/views/admin/ManagePubtator.vue	638
 app/src/views/admin/ManageUser.vue	675
 app/src/views/curate/ApproveReview.vue	841
 app/src/views/curate/ApproveUser.vue	842
-app/src/views/curate/CreateEntity.vue	637
+app/src/views/curate/CreateEntity.vue	540
 app/src/views/curate/ManageReReview.vue	1579
 app/src/views/pages/EntityView.vue	811
 db/11_Rcommands_sysndd_db_table_database_comparisons.R	636
